+++ Contracts/UriInterface.php
@@ -24,6 +24,7 @@ use Stringable;
  *
  * @method string|null getUsername() returns the user component of the URI.
  * @method string|null getPassword() returns the scheme-specific information about how to gain authorization to access the resource.
+ * @method array toComponents() returns an associative array containing all the URI components.
  */
 interface UriInterface extends JsonSerializable, Stringable
 {