+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All Notable changes to `League\Uri\Interfaces` will be documented in this file
 ### Fixed
 
 - Adding Host resolution caching to speed up URI parsing in `UriString`
+- `UriString::parseAuthority` accepts `Stringable` object as valid input
 
 ### Deprecated